Moved DictionaryDataReaderGatewayTest.java and DictionaryFunctionsTest.java to the correct folder structure.
